Completes ex. 10.13-10.15, adds AsyncStorage login
This commit fixes an bug with the previous commit, where the
useSignIn hook did not wait for the login mutation to complete, so
returned the response of the previous login.

This commit also implements an AsyncStorage of authorization token,
and updates the useSignIn hook to place the token into storage on a
successful login.